# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SubsTrack API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SubscriptionBillingCycle =
  (typeof SubscriptionBillingCycle)[keyof typeof SubscriptionBillingCycle];

export const SubscriptionBillingCycle = {
  monthly: "monthly",
  yearly: "yearly",
  weekly: "weekly",
} as const;

export type SubscriptionStatus =
  (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus];

export const SubscriptionStatus = {
  active: "active",
  paused: "paused",
  cancelled: "cancelled",
} as const;

export interface Subscription {
  id: number;
  name: string;
  category: string;
  price: number;
  currency: string;
  billingCycle: SubscriptionBillingCycle;
  startDate: string;
  nextBillingDate: string;
  hasTrial: boolean;
  trialEndDate?: string | null;
  reminderDaysBefore?: number | null;
  notes?: string | null;
  status: SubscriptionStatus;
  createdAt: string;
}

export type SubscriptionCreateBillingCycle =
  (typeof SubscriptionCreateBillingCycle)[keyof typeof SubscriptionCreateBillingCycle];

export const SubscriptionCreateBillingCycle = {
  monthly: "monthly",
  yearly: "yearly",
  weekly: "weekly",
} as const;

export type SubscriptionCreateStatus =
  (typeof SubscriptionCreateStatus)[keyof typeof SubscriptionCreateStatus];

export const SubscriptionCreateStatus = {
  active: "active",
  paused: "paused",
  cancelled: "cancelled",
} as const;

export interface SubscriptionCreate {
  name: string;
  category: string;
  price: number;
  currency?: string;
  billingCycle: SubscriptionCreateBillingCycle;
  startDate: string;
  nextBillingDate: string;
  hasTrial?: boolean;
  trialEndDate?: string | null;
  reminderDaysBefore?: number | null;
  notes?: string | null;
  status?: SubscriptionCreateStatus;
}

export type SubscriptionUpdateBillingCycle =
  (typeof SubscriptionUpdateBillingCycle)[keyof typeof SubscriptionUpdateBillingCycle];

export const SubscriptionUpdateBillingCycle = {
  monthly: "monthly",
  yearly: "yearly",
  weekly: "weekly",
} as const;

export type SubscriptionUpdateStatus =
  (typeof SubscriptionUpdateStatus)[keyof typeof SubscriptionUpdateStatus];

export const SubscriptionUpdateStatus = {
  active: "active",
  paused: "paused",
  cancelled: "cancelled",
} as const;

export interface SubscriptionUpdate {
  name?: string;
  category?: string;
  price?: number;
  currency?: string;
  billingCycle?: SubscriptionUpdateBillingCycle;
  startDate?: string;
  nextBillingDate?: string;
  hasTrial?: boolean;
  trialEndDate?: string | null;
  reminderDaysBefore?: number | null;
  notes?: string | null;
  status?: SubscriptionUpdateStatus;
}

export interface DashboardSummary {
  monthlySpend: number;
  activeCount: number;
  trialsEndingSoon: number;
  upcomingRenewals: number;
  yearlySpend: number;
  savingsOpportunity: number;
}

export interface CategorySpend {
  category: string;
  monthlyAmount: number;
  count: number;
}

export interface MonthlySpend {
  month: string;
  amount: number;
}

export type NotificationAlertType =
  (typeof NotificationAlertType)[keyof typeof NotificationAlertType];

export const NotificationAlertType = {
  trial: "trial",
  renewal: "renewal",
} as const;

export type NotificationAlertSeverity =
  (typeof NotificationAlertSeverity)[keyof typeof NotificationAlertSeverity];

export const NotificationAlertSeverity = {
  info: "info",
  warning: "warning",
  urgent: "urgent",
} as const;

export interface NotificationAlert {
  id: string;
  type: NotificationAlertType;
  title: string;
  message: string;
  daysLeft: number;
  subscriptionId: number;
  subscriptionName: string;
  severity: NotificationAlertSeverity;
}

export type ListSubscriptionsParams = {
  status?: ListSubscriptionsStatus;
  category?: string;
};

export type ListSubscriptionsStatus =
  (typeof ListSubscriptionsStatus)[keyof typeof ListSubscriptionsStatus];

export const ListSubscriptionsStatus = {
  active: "active",
  paused: "paused",
  cancelled: "cancelled",
} as const;