# lib/api-zod/src/generated/types/notificationAlert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SubsTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationAlertSeverity } from "./notificationAlertSeverity";
import type { NotificationAlertType } from "./notificationAlertType";

export interface NotificationAlert {
  id: string;
  type: NotificationAlertType;
  title: string;
  message: string;
  daysLeft: number;
  subscriptionId: number;
  subscriptionName: string;
  severity: NotificationAlertSeverity;
}